# ServerSentEvents: Support CRLF + LF as line endings