--- a/src/SignalR/docs/specs/TransportProtocols.md
+++ b/src/SignalR/docs/specs/TransportProtocols.md
@@ -165,7 +165,7 @@ If the relevant connection has been terminated, a `404 Not Found` status code is
 
 Server-Sent Events (SSE) is a protocol specified by WHATWG at [https://html.spec.whatwg.org/multipage/comms.html#server-sent-events](https://html.spec.whatwg.org/multipage/comms.html#server-sent-events). It is capable of sending data from server to client only, so it must be paired with the HTTP Post transport. It also requires a connection already be established using the `POST [endpoint-base]/negotiate` request.
 
-The protocol is similar to Long Polling in that the client opens a request to an endpoint and leaves it open. The server transmits frames as "events" using the SSE protocol. The protocol encodes a single event as a sequence of key-value pair lines, separated by `:` and using any of `\r\n`, `\n` or `\r` as line-terminators, followed by a final blank line. SignalR currently only supports messages delimmited by `\r\n` and/or `n` due to line buffering. Keys can be duplicated and their values are concatenated with `\n`. So the following represents two events:
+The protocol is similar to Long Polling in that the client opens a request to an endpoint and leaves it open. The server transmits frames as "events" using the SSE protocol. The protocol encodes a single event as a sequence of key-value pair lines, separated by `:` and using any of `\r\n`, `\n` or `\r` as line-terminators, followed by a final blank line. SignalR currently only supports messages delimmited by `\r\n` and/or `\n` due to line buffering. Keys can be duplicated and their values are concatenated with `\n`. So the following represents two events:
 
 ```
 foo: bar