# ServerSentEvents: Support CRLF + LF as line endings

MessageParser now allows a combination of CRLF and LF.

## Description

This PR updates the ServerSentEvent MessageParser to support single LF tokens instead of only a CRLF combination.

This change was made so that different servers libraries that only send `\n` line delimitted messages are supported.

Additionally the documentation should be updated

https://github.com/dotnet/aspnetcore/blob/main/src/SignalR/docs/specs/TransportProtocols.md#server-sent-events-server-to-client-only

> [...] The server transmits frames as "events" using the SSE protocol. The protocol encodes a single event as a sequence of key-value pair lines, separated by : and using any of \r\n, \n or \r as line-terminators, followed by a final blank line.

Other references:

https://html.spec.whatwg.org/multipage/server-sent-events.html#parsing-an-event-stream

> Lines must be separated by either a U+000D CARRIAGE RETURN U+000A LINE FEED (CRLF) character pair, a single U+000A LINE FEED (LF) character, or a single U+000D CARRIAGE RETURN (CR) character.

Fixes #42001

<details>

<summary>
Benchmarks

</summary>

Benchmark System
```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET SDK=8.0.100-preview.7.23360.1
  [Host]     : .NET 8.0.0 (8.0.23.37611), X64 RyuJIT
  Job-YXWEHG : .NET 8.0.0 (8.0.23.35901), X64 RyuJIT

Server=True  Toolchain=.NET Core 8.0  InvocationCount=1
RunStrategy=Throughput  UnrollFactor=1
```


Before:

|             Method |          Input |       Protocol |        Mean |     Error |    StdDev |         Op/s |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |--------------- |--------------- |------------:|----------:|----------:|-------------:|-------:|-------:|------:|----------:|
|  ReadSingleMessage |    NoArguments |           json |    96.62 ns |  1.062 ns |  0.887 ns | 10,350,089.1 | 0.0004 |      - |     - |      72 B |
| WriteSingleMessage |    NoArguments |           json |    45.61 ns |  0.084 ns |  0.074 ns | 21,923,543.6 |      - |      - |     - |         - |
|  ReadSingleMessage |    NoArguments | json-formatted |    95.96 ns |  0.467 ns |  0.414 ns | 10,421,469.1 | 0.0004 |      - |     - |      72 B |
| WriteSingleMessage |    NoArguments | json-formatted |    46.28 ns |  0.471 ns |  0.440 ns | 21,608,991.0 |      - |      - |     - |         - |
|  ReadSingleMessage |   FewArguments |           json |    96.42 ns |  0.406 ns |  0.360 ns | 10,371,281.5 | 0.0005 |      - |     - |      80 B |
| WriteSingleMessage |   FewArguments |           json |    45.32 ns |  0.192 ns |  0.179 ns | 22,067,219.2 |      - |      - |     - |         - |
|  ReadSingleMessage |   FewArguments | json-formatted |   318.40 ns |  1.004 ns |  0.890 ns |  3,140,749.1 | 0.0014 |      - |     - |     288 B |
| WriteSingleMessage |   FewArguments | json-formatted |   524.32 ns |  1.752 ns |  1.553 ns |  1,907,229.0 | 0.0019 |      - |     - |     344 B |
|  ReadSingleMessage |  ManyArguments |           json |   108.89 ns |  0.944 ns |  0.837 ns |  9,183,270.3 | 0.0007 |      - |     - |     128 B |
| WriteSingleMessage |  ManyArguments |           json |    46.30 ns |  0.492 ns |  0.436 ns | 21,596,602.4 |      - |      - |     - |         - |
|  ReadSingleMessage |  ManyArguments | json-formatted | 1,106.27 ns |  6.095 ns |  5.403 ns |    903,936.1 | 0.0038 |      - |     - |     768 B |
| WriteSingleMessage |  ManyArguments | json-formatted | 1,820.49 ns |  7.850 ns |  7.343 ns |    549,303.5 | 0.0038 |      - |     - |     880 B |
|  ReadSingleMessage | LargeArguments |           json | 2,337.62 ns | 10.293 ns |  8.595 ns |    427,785.0 | 0.1259 |      - |     - |  20,560 B |
| WriteSingleMessage | LargeArguments |           json |   350.47 ns |  1.661 ns |  1.473 ns |  2,853,318.8 |      - |      - |     - |         - |
|  ReadSingleMessage | LargeArguments | json-formatted | 4,663.09 ns | 24.086 ns | 21.351 ns |    214,450.0 | 0.2594 |      - |     - |  41,192 B |
| WriteSingleMessage | LargeArguments | json-formatted | 3,878.96 ns | 43.937 ns | 41.099 ns |    257,800.8 | 0.1907 | 0.0076 |     - |  31,304 B |

After:

|             Method |          Input |       Protocol |        Mean |     Error |    StdDev |         Op/s |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |--------------- |--------------- |------------:|----------:|----------:|-------------:|-------:|-------:|------:|----------:|
|  ReadSingleMessage |    NoArguments |           json |   111.80 ns |  0.271 ns |  0.240 ns |  8,944,571.2 | 0.0004 |      - |     - |      72 B |
| WriteSingleMessage |    NoArguments |           json |    45.26 ns |  0.213 ns |  0.199 ns | 22,095,728.8 |      - |      - |     - |         - |
|  ReadSingleMessage |    NoArguments | json-formatted |    99.59 ns |  0.419 ns |  0.392 ns | 10,041,605.8 | 0.0004 |      - |     - |      72 B |
| WriteSingleMessage |    NoArguments | json-formatted |    45.09 ns |  0.081 ns |  0.072 ns | 22,177,353.2 |      - |      - |     - |         - |
|  ReadSingleMessage |   FewArguments |           json |    99.42 ns |  0.442 ns |  0.414 ns | 10,057,961.4 | 0.0005 |      - |     - |      80 B |
| WriteSingleMessage |   FewArguments |           json |    44.96 ns |  0.170 ns |  0.151 ns | 22,240,896.3 |      - |      - |     - |         - |
|  ReadSingleMessage |   FewArguments | json-formatted |   308.42 ns |  2.427 ns |  2.270 ns |  3,242,347.4 | 0.0014 |      - |     - |     288 B |
| WriteSingleMessage |   FewArguments | json-formatted |   527.68 ns |  2.383 ns |  1.990 ns |  1,895,076.9 | 0.0019 |      - |     - |     344 B |
|  ReadSingleMessage |  ManyArguments |           json |   110.79 ns |  0.279 ns |  0.248 ns |  9,026,077.8 | 0.0007 |      - |     - |     128 B |
| WriteSingleMessage |  ManyArguments |           json |    46.60 ns |  0.075 ns |  0.070 ns | 21,457,180.8 |      - |      - |     - |         - |
|  ReadSingleMessage |  ManyArguments | json-formatted |   972.54 ns | 10.774 ns |  9.551 ns |  1,028,234.2 | 0.0038 |      - |     - |     768 B |
| WriteSingleMessage |  ManyArguments | json-formatted | 1,796.23 ns |  7.234 ns |  6.767 ns |    556,721.2 | 0.0038 |      - |     - |     880 B |
|  ReadSingleMessage | LargeArguments |           json | 2,357.61 ns | 15.614 ns | 13.038 ns |    424,158.4 | 0.1259 |      - |     - |  20,560 B |
| WriteSingleMessage | LargeArguments |           json |   355.88 ns |  1.061 ns |  0.886 ns |  2,809,900.6 |      - |      - |     - |         - |
|  ReadSingleMessage | LargeArguments | json-formatted | 4,723.41 ns | 45.398 ns | 40.244 ns |    211,711.6 | 0.2594 |      - |     - |  41,192 B |
| WriteSingleMessage | LargeArguments | json-formatted | 3,903.75 ns | 31.876 ns | 26.618 ns |    256,163.8 | 0.1907 | 0.0076 |     - |  31,304 B |

</details>