@kirides 2023-08-04
Implementation was changed to reflect current decisions.

for now only line buffered formats (`\r\n` & `\n`) are supported.
`\r` is not planned yet, due to complexity and possible delays  
Someone might send `data: hello\r` leaving us having to buffer until we have atleast another byte, worst case two bytes:
`\r` may be followed by ...
1. `\n` for line buffering
2. `\r` meaning current line is complete, wait for next byte to check if we're at end of message, as the second `\r` might just aswell be a part of `\r\n`
3. any other byte, which would mean current line is completed, start next line.