@BrennanConroy 2023-08-04
> 2\. `\r` meaning current line is complete, wait for next byte to check if we're at end of message, as the second `\r` might just aswell be a part of `\r\n`

This sounds almost like `\r` is supported since we're delaying for more bytes 😆 
Now I'm wondering how much extra code it would actually be to support `\r`...