@kirides 2023-08-04
> > 2. `\r` meaning current line is complete, wait for next byte to check if we're at end of message, as the second `\r` might just aswell be a part of `\r\n`
> 
> This sounds almost like `\r` is supported since we're delaying for more bytes 😆 Now I'm wondering how much extra code it would actually be to support `\r`...

With the current proposed implementation, `\r` is "accepted" as valid data if not followed immediately by a `\n`, otherwise it counts towards the CRLF -> EOL.

for a "full" implementation, we would have to handle the following case in some way

```
data: something\r\r
```

if now the server doesn't send any more bytes, how would we know if the `data` is complete?
worst case, we would have to wait until the server closes the connection and accept it as `EOL`.